fix : Wall Werr Wextra

+++ util.hpp
@@ -65,7 +65,7 @@ namespace util
 	int count_semicolon(const std::string & str)
 	{
 		int ret = 0;
-		for (int i = 0; i < str.length(); i++)
+		for (size_t i = 0; i < str.length(); i++)
 		{
 			if (str[i] == ';')
 				ret++;
@@ -76,7 +76,7 @@ namespace util
 	int count_sp(const std::string & str)
 	{
 		int ret = 0;
-		for (int i = 0; i < str.length(); i++)
+		for (size_t i = 0; i < str.length(); i++)
 		{
 			if (str[i] == ' ')
 				ret++;
